Bump dev version and some other changes
